requirements.txt
openai
python-dotenv
pygame
SpeechRecognition
edge-tts
PyAudio